rag/requirements.txt
chromadb>=0.4.22
pypdf>=4.0.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-ruby>=0.20.0
tree-sitter-rust>=0.21.0
tree-sitter-cpp>=0.21.0
tree-sitter-c>=0.21.0
tree-sitter-c-sharp>=0.21.0
openai>=1.12.0
anthropic>=0.18.0
sentence-transformers>=2.3.0
tiktoken>=0.5.2
pyyaml>=6.0.1
rapidfuzz>=3.6.0
numpy>=1.24.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

